requirements.txt
#on MacOS: brew install poppler 

pypdf
pdf2image
pillow
requests
pdfplumber